docker-compose.yml: Docker-compose file to build micros-service application stack of DB, API and web app. Included in here for the first cut is a container that populates the DB with mock clients, if there are none already.
prep-db.sh:
prep-db.js:
dump-db.js:
Dockerfile-client-data:
Define the image for the container that builds the mock data in the DB, for demo purposes.

Dockerfile-client-app:
Image that puts the web app into a container and launches it. This is rough and ready: the whole Angular project is cloned, and run in situ, so it is highly inefficient. I'll do a proper deployment later on but this is enough for a minimal stage of demonstration.